create abstract class for CallActivity and CallNotificationActivity
