+++ app/src/main/res/layout/call_notification_activity.xml
@@ -22,6 +22,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/controllerCallNotificationLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
