+++ app/src/main/res/values/styles.xml
@@ -51,10 +51,6 @@
         <item name="iconTint">@color/fontAppbar</item>
     </style>
 
-    <style name="CallTheme" parent="AppTheme">
-        <item name="android:navigationBarColor">@color/grey950</item>
-    </style>
-
     <style name="BottomNavigationView" parent="@style/Widget.MaterialComponents.BottomNavigationView">
         <item name="elevation">1dp</item>
     </style>
@@ -189,9 +185,9 @@
 
     <!-- Call Launch screen -->
     <style name="AppTheme.CallLauncher">
-        <item name="android:windowBackground">@color/black</item>
-        <item name="android:statusBarColor">@color/black</item>
-        <item name="android:navigationBarColor">@color/black</item>
+        <item name="android:windowBackground">@color/grey950</item>
+        <item name="android:statusBarColor">@color/grey950</item>
+        <item name="android:navigationBarColor">@color/grey950</item>
     </style>
 
     <style name="Nextcloud.Material.TextButton" parent="Widget.MaterialComponents.Button.TextButton.Icon">
